test: re-factor tests and update docstrings
